Crate simplemap

Simple Map with default for missing values and compacting (removal of elements with default value from underlying map).

So you can just:

use simplemap::SimpleMap;

let mut map = SimpleMap::new();

assert_eq!(map[0u32], 0u32);
map[1] = 3;
assert_eq!(map[1], 3);

Structs

SimpleMap
SimpleMap