Struct simplemap::SimpleMap

pub struct SimpleMap<K, V> {
    // some fields omitted
}

SimpleMap

Simple Map with default for missing values and compacting (removal of elements with default value from underlying map).

Methods

impl<K, V> SimpleMap<K, V> where K: Ord + Clone, V: Clone + Eq + Default

fn new() -> SimpleMap<K, V>

Create a SimpleMap.

Default::default() will be used as a default value.

impl<K, V> SimpleMap<K, V> where K: Ord + Clone, V: Clone + Eq

fn new_with_default(default: V) -> SimpleMap<K, V>

Create a SimpleMap with custom default value.

fn compact(&mut self)

Make sure to remove all elements with default value.

fn iter<'a>(&'a self) -> Chain<Iter<'a, K, V>, Map<Iter<'a, (K, V)>, fn(&'a (K, V)) -> (&'a K, &'a V)>>

Iterator over map elements with non-default values.

Note: It might return elements with default value, unless compact is called before iter().

impl<K, V> SimpleMap<K, V> where K: Ord + Clone, V: Clone + Eq

fn iter_cloned<'a>(&'a self) -> Chain<Map<Iter<'a, K, V>, fn((&'a K, &'a V)) -> (K, V)>, Cloned<Iter<'a, (K, V)>>>

Iterator yielding (K, V) instead of (&K, &V)